Separate ACS extract/quantiles logic from columns logic
Until we made the full refactor to convert the OBSColumn thing into
a declarative way, its better to have them apart so we can read
clearly the main methods (Extract/Quantiles)
